LIME-fix
Fixed LIME, where it was initially providing an output for Jupyter Notebook, now saves a .html file in results/explanations

Organized files, analysis of the datasets is now present in results/analysis
The analysis of the performance of predictive models is now in results/model_performance

Streamlined code to reduce lines of codes and simplify structure.
In main.py now it's only required to run init_explainers, instead of LIME and SHAP separately.