More detailed analysis of Default Credit Card clients
With the inconclusive analysis of the dataset, an in-depth analysis was performed to determine what exactly are the existing values in the dataset, to compare with the values present in the dictionary.

As a summary, the pay_amount features (X6-X11) have a considerable amount of records with value -2 (impossible as this value is not defined in the dictionary).
A possible solution is referred in the Jupyter notebook that was added in this commit